Fixed Write-Host Output for $additionalFIles
The foreach loop that copies files from $additionalFiles had a typo with the Write-Host syntax that caused it to fail (looks like there's an issue about it too). I just added an additional $ within the parenthesis to fix it. 

Major props and thanks for creating this guide and these scripts!